initial commit for xAODAnaHelpers python module
