+++ python/utils.py
@@ -1,11 +1,17 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-,
+from __future__ import absolute_import
+from __future__ import print_function
+import logging
+logger = logging.getLogger("xAH.utils")
+logger.setLevel(10) # we use info
+
 from PathResolver import PathResolver
 import random
 
-import logging
-logger = logging.getLogger("xAH_nameGenerator")
-logger.setLevel(10) # we use info
+import os
 
-class xAH_nameGenerator(object):
+class NameGenerator(object):
   adjectives = None
   animals = None
   def __init__(self):
@@ -18,3 +24,10 @@ class xAH_nameGenerator(object):
 
   def __str__(self):
     return self.__repr__()
+
+
+def is_release20():
+  return int(os.environ.get('ROOTCORE_RELEASE_SERIES', 0)) < 25
+
+def is_release21():
+  return not is_release20()
